+++ plugin/definitions.go
@@ -16,3 +16,18 @@ type SignerProvider interface {
 type KeyManager interface {
 	PrivateKey(ctx context.Context, keyID string) (string, error)
 }
+
+// Validator inteface defines the method for signing
+type Validator interface {
+	Verify(ctx context.Context, body []byte, sigtnature []byte) (bool, error)
+}
+
+type ValidatorProvider interface {
+	// initialize a new validator instance with given config
+	New(ctx context.Context, config map[string]string) (Validator, error)
+}
+
+// KeyManager is the interface for key management plugin.
+type PublicKeyManager interface {
+	PublicKey(ctx context.Context, subscriberID string, keyID string) (string, error)
+}
