--- plugin/implementations/verifcation_plugin.go
@@ -1,1 +1,0 @@
-package main
