--- signing_plugin/go.mod
@@ -1,14 +1,0 @@
-module beckn_signing.go
-
-go 1.22.1
-
-require golang.org/x/crypto v0.33.0
-
-require google.golang.org/protobuf v1.36.0 // indirect
-
-require (
-	github.com/tink-crypto/tink-go v0.0.0-20230613075026-d6de17e3f164
-	github.com/tink-crypto/tink-go/v2 v2.3.0
-	golang.org/x/sys v0.30.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1
-)
