--- signing_plugin/plugin_config.yaml
@@ -1,11 +1,0 @@
-plugins:
-  signing_plugin:
-    id: "beckn_signing"
-    config:
-      algorithm: "ed25519"
-      private_key: "private.pem"
-      public_key: "public.pem"
-      request_file: "signing_request.json"
-      verify_file: "verify_request.json"
-      subscriber: "test-subscriber"
-      key_id: "nsdl.co.in.ba|nsdl_bap1|ed25519"
