--- signing_plugin/private.pem
@@ -1,1 +1,0 @@
-XQTrvH1kZippFmizR25yvBD12+5bJCSfUkFw5nFHKnEO7wkaMFk+hHgGVafenaXXKEWBWoWFMJVX8EtRGXUhew==
