--- signing_plugin/public.pem
@@ -1,1 +1,0 @@
-Du8JGjBZPoR4BlWn3p2l1yhFgVqFhTCVV/BLURl1IXs=
