--- signing_plugin/verify_request.json
@@ -1,31 +1,0 @@
-{
-    "method": "POST",
-    "url": "http://localhost:8080",
-    "headers": {
-      "Content-Type": "application/json",
-      "Authorization": "Signature keyId=\"sub123|ukid456|ed25519\",algorithm=\"ed25519\", created=\"1739509253\", expires=\"1739512853\", headers=\"(created) (expires) digest\", signature=\"wu6OBeZDWkYFho5H+1/t11dlB4cIGBeh9Qugm2v4R5URXzAUPOgwg6GThdSih1IoNsTgbv8X2DChYRtFT6h2Cw==\""
-    },
-    "body": {
-      "context": {
-        "domain": "nic2004:60232",
-        "country": "IND",
-        "city": "std:080",
-        "action": "search",
-        "core_version": "1.2.0",
-        "bap_id": "pilot-gateway-1.beckn.nsdl.co.in/option91",
-        "bap_uri": "https://pilot-gateway-1.beckn.nsdl.co.in/option91",
-        "transaction_id": "T1",
-        "message_id": "M1",
-        "timestamp": "2023-06-06T21:00:00.000Z",
-        "ttl": "PT30S"
-      },
-      "message": {
-        "intent": {
-          "category": {
-            "id": "Immediate Delivery"
-          }
-        }
-      }
-    }
-  }
-
